requirements.txt
fastapi
uvicorn
pandas
python-dotenv
sentence-transformers
chromadb
langchain-groq
httpx
fuzzywuzzy
scikit-learn
matplotlib
seaborn
transformers
torch
pydantic
